Use libgit if available and desired
If it is not available, then do not attempt to load and/or build it.
Deal with circular dependency between `magit-git' and `magit-libgit'.
Define the essential function `magit-libgit-repo'.

It's not pretty.

Also see the next commit, which demonstrates how to add a new libgit
based implementation of an existing function.